==> notebooks/4.0-ash-weather-silver-csv-datasets-generator.ipynb <==
# Silver Dataset Generation using Weather Data
<font size=3><strong>Author:</strong> <a href="https://www.linkedin.com/in/~ashkan/" target="_blank">Ashkan Soltanieh</a><br>
<strong>Date:</strong> Jan. 13, 2022</font>

## Table of Contents

<div class="alert alert-success mt-20">
    <ul>
        <li><a href="#Approach">Approach</a></li>
        <li><a href="#Temperature Standardization">Temperature Standardization</a></li>
        <li><a href="#Metadata">Metadata</a></li>
    </ul>
</div>

## Approach:
Bronze dataset has further refined in this notebood. The weather data has been stored as csv dataset and it contains over 121 million rows and 10 columns of hourly data which are already filtered for the location and date of fire recordings. 

Since the data are recorded hourly, and our processing resource is limited, I decided to aggregate hourly data for each day by mean and standard deviation.

Here is the maximum daily standard deviation for the weather elements, generated from bronze dataset:

| Variable  |  Max_Daily_Std   |
| --- | --- |
|t2m  |     10.886982|
|wind_speed | 7.205235|
|cape |    3083.404515|
|d2m  |      11.848188|
|tp   |       0.005628|
|tcc  |       0.509919|
|cvh  |       0.000000|
|cvl  |       0.000000|
|swvl1|       0.151427|

It shows that most of the variables has a very high standard deviation (cosidering their overall range) while high, and low vegitation cover, i.e. **cvh, cvl** have the standard deviation of zero. For these two variables, we will be considering only daily mean aggregation, and for the rest we will be using daily mean and standard deviation aggregation.

## Temperature Standardization
Temperature and dew point temperature are in Kelvin unit. For easier understanding, the value of the data are replaced by Celsius equivalent. It is added into the python script at <code>src/data/weather.py</code> module 

import os
import pandas as pd
import sys
from IPython.display import display

path = os.path.abspath(os.path.join(os.getcwd(), '../data/processed/weather/bronze/bronze_weather-synced-with-fires.csv'))

sys.path.insert(1, os.path.abspath(os.path.join(os.getcwd(),"..","src/data")))
from weather import make_silver_dataframe

dfw = make_silver_dataframe(path)

path_silver = os.path.abspath(
        os.path.join(os.getcwd(), "../data/processed/weather/silver/silver_weather-daily-mean-std.csv"))

dfw.to_csv(path_silver, index = True)

## Metadata

display(dfw.head())
display(dfw.shape)

## <h3 align="center"> Copyright © 2022 - All rights reserved by the author.<h3/>
