==> notebooks/5.0-ash-gold-dataset-generator.ipynb <==
# Gold Dataset Generator
<font size=3><strong>Author:</strong> <a href="https://www.linkedin.com/in/~ashkan/" target="_blank">Ashkan Soltanieh</a><br>
<strong>Date:</strong> Jan. 15, 2022</font>

## Table of Contents

<div class="alert alert-success mt-20">
    <ul>
        <li><a href="#Approach">Approach</a></li>
        <li><a href="#Merging Area of Burn and Characteristics Data">Merging Area of Burn and Characteristics Data</a></li>
        <li><a href="#Duplication Check in Merged Data">Duplication Check in Merged Data</a></li>
        <li><a href="#Dealing with Duplicates">Dealing with Duplicates</a></li>
        <li><a href="#Quartile Analysis">Quartile Analysis</a></li>
        <li><a href="#Metadata">Metadata</a></li>
    </ul>
</div>

## Approach:
So far we have merged and cleaned part of the data. As a quick overview, we have cleaned the two datasets for wildfires, as well as merged and cleaned weather data. Additionally, weather data are aggregated by their daily mean and standard deviation to bring consistency into dataset and alignment with wildfire datasets. The process of cleaning the weather data to create the focus weather dataset based on wildfire date and location and drop the rest of redundancy is completed, and silver dataset contains required aggregated variables for merging datasets.

In this notebook, our goal is to merge all dataset and start data preprocessing and exploratory analysis. The keys for merging Area of Burn and characteristics datasets are UID_Fire, and REF_ID, and Date_of_Burn. For merging wildfire and weather data we will be using rounded spatial locations data and date.

Finally, we are going to categorize the fire data into five categories. The categories will be selected based on quartile analysis for area of burn data. Ranking wildfire based on different citeria is a common practice. Here is an example of this has been done in Government of BC ([Wildfire Ranking](https://www2.gov.bc.ca/gov/content/safety/wildfire-status/about-bcws/wildfire-response/fire-characteristics/rank)).

import os
import numpy as np
import pandas as pd
import sys
from IPython.display import display
sys.path.insert(1, os.path.abspath(os.path.join(os.getcwd(),"..","src/data")))
from weather import get_rounded_locations

path_weather = os.path.abspath(os.path.join(os.getcwd(), '../data/processed/weather/silver/silver_weather-daily-mean-std.csv'))
path_aob = os.path.abspath(os.path.join(os.getcwd(), '../data/processed/wildfire/silver/silver_AoB.csv'))
path_characteristics = os.path.abspath(os.path.join(os.getcwd(), '../data/processed/wildfire/silver/silver_chracteristics.csv'))

df_weather = pd.read_csv(path_weather)
df_aob = pd.read_csv(path_aob)
df_characteristics = pd.read_csv(path_characteristics)

## Merging Area of Burn and Characteristics Data

df_aob_merge = df_aob.rename(columns={'Date_of_Burn' : 'Date'}).set_index(['UID_Fire', 'REF_ID', 'Date'])
df_characteristics_merge = df_characteristics.set_index(['UID_Fire', 'REF_ID', 'Date'])
df_characteristics_aob = df_characteristics_merge.merge(df_aob_merge, on=['UID_Fire', 'REF_ID', 'Date'], how='inner')
df_characteristics_aob.reset_index(drop=False, inplace=True)
df_characteristics_aob.head(3)

## Duplication Check in Merged Data
To check duplication after merging data we cannot use <code>rounded_lat</code> and <code>rounded_lon</code> data as the original area of burn data are not uniquely identified with any geospatioal data. On the other hand, Fire Status and Satlite data are needed in classification in order to distinguish the status type of the fire and which satelite is being used for detecting fire characteristics.

df_temp = df_characteristics_aob.set_index(['UID_Fire', 'REF_ID', 'Date', 'Total_AoB', 'Status', 'sat'])
display('Duplicated fire data after merging:', np.unique(df_temp.index.duplicated(), return_counts=True))

## Dealing with Duplicates
Observing duplicated indices using <code>['UID_Fire', 'REF_ID', 'Date', 'Total_AoB', 'Status', 'sat']</code> shows that 9,614 number of the records have different rounded latitude and longitude (i.e. belonging to different quater geospatial degree grouping), while referring to the same fire incident. As a result to have more trust-worthy dataset we need to only keep one of the duplicated records and drop the rest. We will be keeping first observation of the duplicated records.

df_characteristics_aob = df_characteristics_aob.groupby(['UID_Fire', 'REF_ID', 'Date', 'Total_AoB', 'Status', 'sat'], 
                                                        axis=0, 
                                                        as_index=False).first()

As expected, no duplication observed after merging in the data.

display(df_aob.shape, df_aob.head(3))
display(df_characteristics.shape, df_characteristics.head(3))
display(df_characteristics_aob.shape, df_characteristics_aob.head(3))

Comparison between characteristics and area of burn data shows that 14,184 fire observations have a corresponding area of burn recording. These data will be further refined during exploratory analysis.

Here we're just dropping rounded_lat, and rounded_lon data as they're just creating for merging weather and fire tables

df_characteristics_aob_merge = df_characteristics_aob.set_index(['rounded_lat', 'rounded_lon', 'Date'])
df_weather_merge = df_weather.rename(columns = {'latitude': 'rounded_lat', 'longitude': 'rounded_lon', 'date':'Date'}).set_index(['rounded_lat', 'rounded_lon', 'Date'])
df_characteristics_aob_weather = df_characteristics_aob_merge.merge(df_weather_merge, on=['rounded_lat', 'rounded_lon', 'Date'], how='inner')
df_characteristics_aob_weather.reset_index(drop=False, inplace=True)
df_characteristics_aob_weather.set_index(['UID_Fire', 'REF_ID', 'Date'], inplace=True)
df_characteristics_aob_weather.drop(['rounded_lat', 'rounded_lon'], axis=1, inplace=True)

df_characteristics_aob_weather.columns

## Quartile Analysis

We couldn't find any categorizing only based on wildfire burn area from the available references and litrature review. Therefore, since the current fire observations are high enough we decided to do fair distribution into five categories using quartile analysis based on quantile limits of Total_AoB data.

Q1 = df_characteristics_aob_weather['Total_AoB'].quantile(0.25)
median = df_characteristics_aob_weather['Total_AoB'].quantile(0.5)
Q3 = df_characteristics_aob_weather['Total_AoB'].quantile(0.75)
lower_extreme = Q1 - 1.5 * (Q3 - Q1) if Q1 - 1.5 * (Q3 - Q1) > 0 else 0 
upper_extreme = Q3 + 1.5 * (Q3 - Q1)

display(f"Q1: {Q1}", 
        f"median: {median}", 
        f"Q3: {Q3}", 
        f"lower_extreme: {lower_extreme}", 
        f"upper_extreme: {upper_extreme}")

df_characteristics_aob_weather["AoB_Category"] = ''

df_characteristics_aob_weather.loc[
    (df_characteristics_aob_weather['Total_AoB'] > lower_extreme) &
    (df_characteristics_aob_weather['Total_AoB'] <= Q1), 'AoB_Category'] = 'Very Low'

df_characteristics_aob_weather.loc[
    (df_characteristics_aob_weather['Total_AoB'] > Q1) &
    (df_characteristics_aob_weather['Total_AoB'] <= median), 'AoB_Category'] = 'Low'

df_characteristics_aob_weather.loc[
    (df_characteristics_aob_weather['Total_AoB'] > median) &
    (df_characteristics_aob_weather['Total_AoB'] <= Q3), 'AoB_Category'] = 'Moderate'

df_characteristics_aob_weather.loc[
    (df_characteristics_aob_weather['Total_AoB'] > Q3) &
    (df_characteristics_aob_weather['Total_AoB'] <= upper_extreme), 'AoB_Category'] = 'High'

df_characteristics_aob_weather.loc[
    (df_characteristics_aob_weather['Total_AoB'] > upper_extreme), 'AoB_Category'] = 'Very High'

df_characteristics_aob_weather['AoB_Category'].value_counts()

df_characteristics_aob_weather.columns = ['total_aob', 'fire_status', 'satelite', 'lat_mean', 'lat_std',
       'lon_mean', 'lon_std', 't21_mean', 't21_std', 't31_mean', 't31_std', 'frp_mean', 'frp_std', 'conf_mean', 'conf_std', 'temp_mean', 'temp_std', 'convective_energy_mean', 'convective_energy_std', 'dewpoint_temp_mean',
       'dewpoint_temp_std', 'total_precipitation_mean', 'total_precipitation_std', 'total_cloud_cover_mean', 
       'total_cloud_cover_std', 'high_veg_cover_mean', 'low_veg_cover_mean', 'soil_water_mean', 'soil_water_std', 
       'wind_speed_mean', 'wind_speed_std', 'category_aob']

## Metadata

display(df_characteristics_aob_weather.head())
display(df_characteristics_aob_weather.shape)

path_gold = os.path.abspath(
        os.path.join(os.getcwd(), "../data/processed/gold.csv"))
df_characteristics_aob_weather.to_csv(path_gold, index = True)

## <h3 align="center"> Copyright © 2022 - All rights reserved by the author.<h3/>
